added review

+++ script.js
@@ -26,56 +26,35 @@ const target = event.target;*/
 
   //check if clicked element is a button & if not, exit from function
 
-  if (!target.matches("button")) {
-    return;
-  }
+  if (!target.matches("button")) return;
+  
+  if (target.classList.contains("operator")) handleOperator(target.value); // formatting -> less lines
+  
+  if (target.classList.contains("buttonDot")) inputDecimal(target.value);
+  
+  if (target.classList.contains("all-clear")) resetCalculator();
+  
+  if (target.classList.contains("delete")) deletePrevious();
+  
+  if(!isNaN(Number.parseInt(target.value))) inputDigit(target.value);
 
-  if (target.classList.contains("operator")) {
-    handleOperator(target.value);
-    updateDisplay();
-    return;
-  }
-
-  if (target.classList.contains("buttonDot")) {
-    inputDecimal(target.value);
-    updateDisplay();
-    return;
-  }
-
-  if (target.classList.contains("all-clear")) {
-    resetCalculator();
-    updateDisplay();
-    return;
-  }
-
-  if (target.classList.contains("delete")) {
-    deletePrevious();
-    updateDisplay();
-    return;
-  }
-
-  inputDigit(target.value);
-  updateDisplay();
+  updateDisplay(); // always needs to be called
 });
 
 function inputDigit(digit) {
   const { displayValue, waitingForSecondOperand } = calculator;
   // Overwrite 'displayValue' if the current value is "o"
-  if (waitingForSecondOperand === true) {
-    calculator.displayValue = digit;
-    calculator.waitingForSecondOperand = false;
-  } else {
-    calculator.displayValue =
-      displayValue === "0" ? digit : displayValue + digit;
-  }
+  if(!waitingForSecondOperand) return calculator.displayValue = displayValue === "0" ? digit : displayValue + digit; // same as if(waitingForSecondOperand === false) BUT carefull, will also be true if value is 0, null or undefined
+  
+  calculator.displayValue = digit;
+  calculator.waitingForSecondOperand = false;
 }
 
 function inputDecimal(dot) {
   //if the "displayValue" property does not contain a decimal point
-  if (!calculator.displayValue.includes(dot)) {
-    //append the decimal point
-    calculator.displayValue += dot;
-  }
+  if (calculator.displayValue.includes(dot)) return;
+
+  calculator.displayValue += dot;
 }
 
 function handleOperator(nextOperator) {
@@ -84,10 +63,7 @@ function handleOperator(nextOperator) {
   //"parseFloat" converts the string contents of 'displayValue' to a floating-point number
   const inputValue = parseFloat(displayValue);
 
-  if (operator && calculator.waitingForSecondOperand) {
-    calculator.operator = nextOperator;
-    return;
-  }
+  if (operator && calculator.waitingForSecondOperand) return calculator.operator = nextOperator; //formatting -> less lines
 
   //verify that 'firstOperand' is null and that the "inputValue" is not a 'nan' value
   if (firstOperand === null && !isNaN(inputValue)) {
@@ -105,21 +81,16 @@ function handleOperator(nextOperator) {
 }
 
 function calculate(firstOperand, secondOperand, operator) {
-  if (operator === "+") {
-    return firstOperand + secondOperand;
-  } else if (operator === "-") {
-    return firstOperand - secondOperand;
-  } else if (operator === "*") {
-    return firstOperand * secondOperand;
-  } else if (operator === "/") {
-    return firstOperand / secondOperand;
-  }
-
+  if (operator === "+") return firstOperand + secondOperand; // we can omit the else because of the early return statement + formatting -> less lines
+  if (operator === "-") return firstOperand - secondOperand;
+  if (operator === "*") return firstOperand * secondOperand;
+  if (operator === "/") return firstOperand / secondOperand;
+  
   return secondOperand;
 }
 
 function deletePrevious() {
-  if ("delete") {
+  if ("delete") { // not sure about this if check, this will always be true?
     return (calculator.displayValue = calculator.displayValue.slice(0, -1));
   }
 }
@@ -129,4 +100,5 @@ function resetCalculator() {
   calculator.firstOperand = null;
   calculator.waitingForSecondOperand = false;
   calculator.operator = null;
+  updateDisplay(); // update display value was forgotten
 }
